requirements.txt
django
tzdata
django-cities-light
python-dotenv
djangorestframework==3.14.0
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
